+++ test/utils/index.js
@@ -1,0 +1,11 @@
+const defaultGasLimit = '0x100000';
+
+const toConfirmationPromise = (promiEvent) => new Promise(
+  (resolve, reject) => promiEvent.on('confirmation',
+    (confirmationNumber, receipt) => resolve(receipt)).catch(reject),
+);
+
+module.exports = {
+  defaultGasLimit,
+  toConfirmationPromise
+}
